+++ config/template.go
@@ -1,9 +1,0 @@
-package config
-
-import "text/template"
-
-var TPL *template.Template
-
-func init() {
-	TPL = template.Must(template.ParseGlob("templates/*.gohtml"))
-}