+++ model/userHandler.go
@@ -3,14 +3,17 @@ package model
 import (
 	"net/http"
 
-	"github.com/nstoker/gorocktrack/config"
+	"github.com/nstoker/gorocktrack/app"
+	"github.com/sirupsen/logrus"
 )
 
 func UserIndex(w http.ResponseWriter, r *http.Request) {
 	users, err := ShowAllUsers()
 	if err != nil {
-		http.Error(w, http.StatusText(500), http.StatusInternalServerError)
+		logrus.Errorf("UserIndex: %s", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-	config.TPL.ExecuteTemplate(w, "users.index.gohtml", users)
+
+	app.TPL.ExecuteTemplate(w, "users_index.gohtml", users)
 }