+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rebooted to simplify structure and get some basics right before too deep.
 - Following [a tutorial](https://kb.objectrocket.com/category/postgresql?filter=Go+Lang+and+PostgreSQL+Web+App+MVC+pattern+Part).
   - Part 2 completed.
+- Table migration configured.
+- Basic user model and seeder created.
+
+### ERRORS
+
+1. The password should be encrypted. So...
 
 ## v0.0.1 2020-10-17